+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 9.4.1
 
+* Replace warning about two bots from `LongPolling` to `DefaultKtorRequestsExecutor`
+
 ## 9.4.0
 
 * `Version`: